requirements.txt
python-dotenv~=0.21.0
autotech-sdk==1.0.2
dacite==1.6.0